Create mgzsev_data.csv
MG ZSEV PIDs. 
The true SoC on the MGZSEV depends on the BMS software standard in the car, so the SoC_Raw is factored to make it read correctly 0 to 100%.  This correction is done automatically in Torque Pro based on the BMS software standard, with SoC_Display being the corrected value.  I can keep this file updated if you decide to use it.