Added PID json file for MiniCooperSE electric
The protocol "B" is unsure, could be 6 or 7 instead. "B" is entered because of the AT command "ATSPB" in the initialisation commands

The equations Letters are also unsure : example : received data for 22DDBC (soc) to module 607 is : 607F1100962DDBC01A6607F12103E50091FFFF 
that is interpreted by me as :
607 F1 10 09 62 DD BC 01 A6 607 F1 21 03 E5 00 91 FF FF
?      A  B  C    D   E    F    G   H    ?     I    J   K   L   M   N  O  P
the actual SOC is found in the 01 A6 pair. In my interpretation that is G + H : 01A6 in hex equals 422 in decimal. Divide by 10 to get the SOC : 42.2%
therefore the equation is (G*256+h)*0.1
if the 01 A6 pair is actually byte A and B the equation must be altered to (A*256+B)*0.1