Reduced timeout from 1024 mSec to
Setting ATSTFF to ATST0F. ST is the timeout command. FF is a timeout period of 1024 msec which is far too long, the message comes almost instantly. setting the value to 0F instead of FF brings the timeout down to 64 msec which works fine (tested in my own iOs app). ATST00 also works but is slower. ATST01 (4 msec timeout is too short, no data is received)